refactor WeaponType package
